+++ src/openai_python_api/models.py
@@ -20,7 +20,8 @@ COMPLETIONS = (
     "gpt-4-32k",
     "gpt-4-32k-0613",
     # GPT-3
-    "gpt-3.5-turbo-1106" "gpt-3.5-turbo",
+    "gpt-3.5-turbo-1106",
+    "gpt-3.5-turbo",
     "gpt-3.5-turbo-16k",
     # DEPRECATION WARNING: GPT-3
     "gpt-3.5-turbo-0613",
